Fix workgroup select for import and new borehole
